training/requirements.txt
torch>=2.0
torchvision>=0.15
albumentations>=1.3
python-chess>=1.9
Pillow>=10.0
cairosvg>=2.7
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.65
onnx>=1.14
ai-edge-torch>=0.2; python_version < "3.13"
